club_world_cup_bot/requirements.txt
aiogram>=3.0.0
python-dotenv==1.0.0
APScheduler==3.10.4
requests==2.31.0 